backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
psycopg2-binary==2.9.6
pytest==7.4.0
pytest-flask==1.2.0
requests==2.31.0
werkzeug==2.3.6  